Reworked acc/dec + Reworked movement state
Movement system, as well as movement states, now contain a groundAcc and airAcc (same for decel) instead of having air vs total.
Added a (configurable) mechanic to grappling where less ground decel is applied as the player climbs above their max normal speed, for more forgiveness for grazing the ground

After that, I reworked the movement state system to have a type: Base or Modifier. Base movementstates include default, dash, post dash, and grapple, and work by taking only the effects of the highest priority one. Modifiers will, in the future, multiply/modify the values of the current base one, such as having a speed debuff or a no-jump debuff.

This fixed the issue with the dash state conflicting with grappling, as well. The grapple simply takes priority now.