Created robust and expandable enemy ai system
Created generic Enemy AI Controller that loads in an Enemy Config scriptable object. That scriptable object, for right now, contains ability to pass in an initial AIBehavior state and the ability to create StateTransitions (which take a starting state, ending state, and AICondtion).

The AI Controller works by checking each of its StateTransitions. If we are on the starting state, check the condition. If the condition returns true, execute the transition to the next state.

The states themselves have onactivate and ondeactivate for transition code, and have tick and physics tick to determine behaviors during that state.